refactoring frontend with auth
