+++ intercom/src/components/Navigation/Navigation.js
@@ -1,87 +1,12 @@
 import React from 'react';
 import { Link, NavLink } from 'react-router-dom';
 
-// Firebase App (the core Firebase SDK) is always required and
-// must be listed before other Firebase SDKs
-import firebaseConfig from './firebaseConfig';
-
 // data management
 import { useStateValue } from 'react-conflux';
 import { globalContext } from '../../store/contexts';
-import { SET_USER } from '../../store/constants';
 
-//firebase
-let firebase = require('firebase/app');
-require('firebase/auth');
-
-firebase.initializeApp(firebaseConfig);
-let provider = new firebase.auth.GoogleAuthProvider();
-firebase.auth().useDeviceLanguage();
-
-const Navigation = () => {
+const Navigation = props => {
     const [state, dispatch] = useStateValue(globalContext);
-
-    const handleLogin = () => {
-        firebase
-            .auth()
-            .signInWithPopup(provider)
-            .then(function(result) {
-                let token = result.user.getIdToken();
-                console.log(token);
-                dispatch({ type: SET_USER, payload: { token: token } });
-                // dispatch to store to save the token for later use
-                return token;
-            })
-            .then(token => {
-                // Need to get data from database with this fetch/axios call
-                const url = 'http://localhost:3300';
-                fetch(url, {
-                    method: 'GET', // or 'PUT'
-                    headers: {
-                        'Content-Type': 'application/json',
-                        application: token
-                    }
-                })
-                    .then(response =>
-                        // Once fetch is done store the data via dispatch to the store
-                        // dispatch({ type: SET_USER, payload: userProfile });
-                        console.log('Success:', JSON.stringify(response))
-                    )
-                    .catch(error => console.error('Error:', error));
-            })
-            .catch(function(error) {
-                // Handle Errors here.
-                let errorCode = error.code;
-                let errorMessage = error.message;
-                // The email of the user's account used.
-                let email = error.email;
-                // The firebase.auth.AuthCredential type that was used.
-                let credential = error.credential;
-                // ...
-            });
-    };
-
-    const handleLogout = () => {
-        firebase
-            .auth()
-            .signOut()
-            .then(function() {
-                // Sign-out successful.
-                let userProfile = {
-                    uid: '',
-                    email: '',
-                    photo: '',
-                    name: ''
-                };
-                dispatch({ type: SET_USER, payload: userProfile });
-                // history.push('/');
-                console.log('Signout success!');
-            })
-            .catch(function(error) {
-                // An error happened.
-                console.error('Signout Error', error);
-            });
-    };
 
     return (
         <nav className="navbar navbar-inverse navbar-fixed-top navbar-container">
@@ -144,7 +69,7 @@ const Navigation = () => {
                         <a href="#contact">Contact</a>
                     </li>
                     <li>
-                        <p onClick={handleLogin}>Login</p>
+                        <p onClick={props.handleLogin}>Login</p>
                     </li>
                     <li>
                         {/* <Link
