+++ intercom/src/firebaseConfig.js
